refactor: rename ProjectPage.tsx to ProjectDetailView.tsx for clarity and consistency
