+++ apps/web/src/features/projects/views/ProjectDetailView.tsx
